Train stereo.
