+++ examples/stereo_tracking_example/run.py
@@ -2,7 +2,7 @@ import configparser
 import os
 import sys
 
-sys.path.append('../')
+sys.path.append('../../')
 
 from typing import List
 import torch
@@ -61,7 +61,7 @@ def get_videofile_reader(path: str, name: str) -> VideoReader:
 
 
 if __name__ == '__main__':
-    checkpoint = torch.load(os.path.join(os.path.dirname(__file__), '..', 'tests', 'test_data', 'best.ckpt'),
+    checkpoint = torch.load(os.path.join(os.path.dirname(__file__), '../..', 'tests', 'test_data', 'best.ckpt'),
                             map_location=torch.device('cpu'))['model']
     model = mobilestereonet(checkpoint)
     # checkpoint = torch.load(os.path.join(os.path.dirname(__file__), '..', 'tests', 'test_data', 'best.pth.tar'),
@@ -81,7 +81,7 @@ if __name__ == '__main__':
         readers.append(get_videofile_reader(file_src, os.path.basename(file_src)))
 
     config = configparser.ConfigParser()
-    config.read(os.path.join(os.path.dirname(__file__), '..', 'tests', 'test_data', 'conf.txt'))
+    config.read(os.path.join(os.path.dirname(__file__), '../..', 'tests', 'test_data', 'conf.txt'))
     bboxes_list = eval(config.get('bboxes', 'values'))
     bboxes = []
     for bb in bboxes_list:
