+++ models/models.py
@@ -6,6 +6,8 @@ from models.defects.blocks import OutBlock
 from models.defects.ssd300 import SSD300
 from models.defects.vgg19 import InputRescale
 from models.preprocessing import *
+from models.stereo.blocks import GANetInputBlock, GANetBackbone, GANetOutputBlock, DepthOutput, MobileStereoNetBackbone, \
+    MobileStereoNetInputBlock
 
 
 class ModelBuilder(nn.Module):
@@ -147,10 +149,10 @@ def defects_model(path_to_templates=None, emb_size: int = 512, weights=None, is_
         output_block=OutBlock(-1, -1).to(device)
     )
 
-def mobilestereonet(weights = None):
+def mobilestereonet(weights = None, is_train=False, device='cuda'):
     input_block = MobileStereoNetInputBlock()
-    backbone = MobileStereoNetBackbone()
-    output_block = DepthOutput()
+    backbone = MobileStereoNetBackbone(training=is_train)
+    output_block = DepthOutput(training=is_train)
     if weights:
         weight_index = input_block.import_weights(weights)
         weight_index = backbone.import_weights(weights, weight_index)
@@ -178,3 +180,4 @@ def ganet(weights = None):
     )
     torch.nn.DataParallel(model)
     return model
+
